requirements.txt
torch
numpy
pandas
pillow
tqdm
primefac
scipy
texttable